v1.9
You can now add as many folders as you want, each of which will be recursively searched for media-files that are referenced in the movie's XML project file (.wlmp). Folders can be added by drag-drop or by "Button 2". A new "Reset" item appears in the main menu that will clear the list of media-file folders.